updated runner script, requirements
