--- requirements.yml
@@ -1,50 +1,0 @@
-name: actogram
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - ca-certificates=2020.6.20
-  - certifi=2020.6.20
-  - cycler=0.10.0
-  - freetype=2.10.2
-  - jpeg=9d
-  - kiwisolver=1.2.0
-  - lcms2=2.11
-  - libblas=3.8.0
-  - libcblas=3.8.0
-  - libcxx=10.0.1
-  - libffi=3.2.1
-  - libgfortran=4.0.0
-  - liblapack=3.8.0
-  - libopenblas=0.3.10
-  - libpng=1.6.37
-  - libtiff=4.1.0
-  - libwebp-base=1.1.0
-  - llvm-openmp=10.0.1
-  - lz4-c=1.9.2
-  - matplotlib=3.3.0
-  - matplotlib-base=3.3.0
-  - ncurses=6.2
-  - numpy=1.19.1
-  - olefile=0.46
-  - openssl=1.1.1g
-  - pandas=1.1.0
-  - pillow=7.2.0
-  - pip=20.2.1
-  - pyparsing=2.4.7
-  - python=3.7.8
-  - python-dateutil=2.8.1
-  - python_abi=3.7
-  - pytz=2020.1
-  - readline=8.0
-  - scipy=1.5.2
-  - setuptools=49.2.1
-  - six=1.15.0
-  - sqlite=3.32.3
-  - tk=8.6.10
-  - tornado=6.0.4
-  - wheel=0.34.2
-  - xz=5.2.5
-  - zlib=1.2.11
-  - zstd=1.4.5
-prefix: /opt/anaconda3/envs/actogram
